move search routes to flutter api
