+++ mongoClient/mongo_routes.py
@@ -1,4 +1,3 @@
-import os
 from bson import ObjectId
 from flask import Blueprint, jsonify, request
 import pymongo
@@ -83,45 +82,3 @@ def delete_data(id):
     except Exception as e:
         return f"An error occurred: {e}", 400
     
-@mongo_bp.route('/search_generic', methods=['GET'])
-def search_generic():
-    try:
-        collection = mongoClient.get_collection(db="grocerydb", collection="genericItems")
-            
-        query = request.args.get("query")
-        if not query:
-            return jsonify({"error": "Query parameter is required"}), 400
-        
-        agg_pipeline = [
-            {
-                "$search": {
-                    "index": "default",
-                    "autocomplete": {
-                        "query": query,
-                        "path": "genericItem",
-                        "tokenOrder": "any",
-                        "fuzzy": {
-                            "maxEdits": 2,
-                            "prefixLength": 1,
-                            "maxExpansions": 256
-                        }
-                    }
-                }
-            },
-            {"$limit": 20}
-        ]
-
-        results = list(collection.aggregate(agg_pipeline))
-        print('results:', results)
-
-        # Convert ObjectId fields to strings
-        for doc in results:
-            doc["_id"] = str(doc["_id"])
-
-        return jsonify(results), 200
-    except Exception as e:
-        return f"An error occurred: {e}", 400
-    
-@mongo_bp.route('/search_store', methods=['GET'])
-def search_store():
-    pass
